+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 - Safer check dataframe series is of integer dtype (#179)
 
 ### Added
-- Possibility to choose the table name for the report (#183)
-- Can set all `spatialdata_io.xenium` arguments in `sopa.io.xenium`
 - Can filter low-quality transcript for transcript-based segmentation (#79)
+- Possibility to choose the table name for the report (#183)
+- Possibility to choose the table name for `sopa.io.explorer.write` (#183)
+- Can set all `spatialdata_io.xenium` arguments in `sopa.io.xenium`
 
 ## [2.0.0] - 2025-01-20